requirements.txt
google-cloud-firestore
Flask
gunicorn
nltk
